feat(docker): Reproducible `cannon` prestate
## Overview

Adds a dockerfile for reproducible builds of the `cannon` prestate with `kona-client`.

Note that `cannon` does not yet support `kona`. https://github.com/ethereum-optimism/optimism/pull/14454 must be merged before this can be used. This PR just adds the route to build the artifacts, but we will need to follow up with <issue_num> once the new instructions are supported and a new `cannon` tag has been released.