Set bootnodes for all superchains
This ensures that bootnodes are set for all superchains. This uses
the parent superchain to determine if it is a testnet or a mainnet.
This fixes a bug where for any chain besides OPM or Base, bootnodes
would not be set by default.

This also fixes the tests broken in the last commit.