Only check that the listen port & IP is available
`check_ports_inner` binds both a UDP and TCP socket to the supplied IP & PORT.
This is invalid for binding to the advertised IP because the advertised IP is
usually an external IP (either or a load balancer or the user's router when
behind NAT at home or the machine's IP when running in docker) rather than an
IP that the kernel is able to bind to. If the IP is not the same as the local
IP of the context in which kona-node is running, OS Error 99: Cannot assign
requested address is returned.