Splitting build and seal into separate engine tasks
The main goal of this change is to allow engine processing to
continue after block building has been initiated but before
it has been sealed.

This is accomplished by
- Updating the EngineActor to
  - Make the BuildTask only initiate block building
  - Add a new SealTask to seal and insert the initiated block
- Updating the SequencerActor loop to seal the last block and
start the next block. Importantly the build_block function no
longer starts and completes block building in a single call but
rather seals the last block if there was one, and starts the next.
This means that the time spent waiting for the next tick will be
_during block building_.