Wraps EngineClient struct in a trait to allow mocking in task tests
The EngineClient struct is a dependency of all EngineActor tasks.
This creates an EngineClient trait, moving the EngineClient struct
to the OpEngineClient name, and makes all tasks generic over this
trait.

This is a larger change since generics are infectious, and all of
the references up to and through EngineActor also needed to become
generic.