refactor(node/service): refactor the `L1WatcherRpc` actor
The actor contained two event loops that are now joined in one in
`L1WatcherActor`.  The related `Context` and `State` structs are merged
into the primary struct.  The `RootProvider` field has been changed to a
`Provider` trait to improve testability.

Further, the `BlockStream` type has been isolated and rewritten to not
use a reference to a `RootProvider` but a `Provider` trait.