feat(node/EngineActor): Wraps `EngineClient` in a trait to allow mocking in task tests (#3123)
The EngineClient struct is a dependency of all `EngineActor` tasks and
makes it extremely hard to unit test most parts of the `EngineActor`
without running a dummy server (which would introduce many different
ways for unit tests to fail that are not related to the code under
test).

This creates an `EngineClient` trait, moving the `EngineClient` struct
to `OpEngineClient`, and makes all tasks generic over this trait.

This is a larger change since generics are infectious, and all of the
references up to and through `EngineActor` also needed to become
generic.

Also creates mocks and examples of how to use this client in testing
(see: `BuildTask` tests).

ref: #3073 since this is one step in making the EngineActor unit
testable